restructured ds-setup project
