+++ config-overrides.js
@@ -6,19 +6,10 @@ module.exports = function override(config) {
     process: "process/browser",
     stream: require.resolve("stream-browserify"),
     vm: require.resolve("vm-browserify"),
-    // stream: "stream-browserify",
-    // process: require.resolve("process/browser"),
-    // stream: require.resolve("stream-browserify"),
-    // assert: require.resolve("assert"),
-    // http: require.resolve("stream-http"),
-    // https: require.resolve("https-browserify"),
-    // os: require.resolve("os-browserify"),
-    // url: require.resolve("url"),
   });
   config.resolve.fallback = fallback;
   config.plugins = (config.plugins || []).concat([
     new webpack.ProvidePlugin({
-      //process: "process/browser",
       Buffer: ["buffer", "Buffer"],
     }),
   ]);
