+++ package.json
@@ -8,9 +8,6 @@
     "@dm3-org/dm3-lib-profile": "^1.3.0",
     "@rainbow-me/rainbowkit": "^2.1.1",
     "@tanstack/react-query": "^5.28.4",
-    "@testing-library/jest-dom": "^6.2.0",
-    "@testing-library/react": "^14.1.2",
-    "@testing-library/user-event": "^14.5.2",
     "@types/jest": "^29.5.2",
     "@types/node": "^18.19.3",
     "@types/react": "^18.3.0",
@@ -58,6 +55,9 @@
     "crypto-browserify": "^3.12.0",
     "react-app-rewired": "^2.2.1",
     "stream-browserify": "^3.0.0",
-    "vm-browserify": "^1.1.2"
+    "vm-browserify": "^1.1.2",
+    "@testing-library/jest-dom": "^6.2.0",
+    "@testing-library/react": "^14.1.2",
+    "@testing-library/user-event": "^14.5.2"
   }
 }
