+++ src/index.css
@@ -6,3 +6,9 @@ body {
   -webkit-font-smoothing: antialiased;
   -moz-osx-font-smoothing: grayscale;
 }
+
+.main-container {
+  display: flex;
+  justify-content: flex-end;
+  padding: 12,
+}
