+++ src/index.tsx
@@ -1,18 +1,17 @@
-import { getDefaultConfig, RainbowKitProvider } from "@rainbow-me/rainbowkit";
-import "@rainbow-me/rainbowkit/styles.css";
-import React from "react";
-import ReactDOM from "react-dom/client";
-import { WagmiProvider } from "wagmi";
-import { mainnet, optimism, sepolia } from "wagmi/chains";
 import "./index.css";
 import "./polyfills";
-
+import React from "react";
+import App from "./components/App";
+import { WagmiProvider } from "wagmi";
+import ReactDOM from "react-dom/client";
+import "@rainbow-me/rainbowkit/styles.css";
+import { mainnet, optimism, sepolia } from "wagmi/chains";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import App from "./App";
+import { getDefaultConfig, RainbowKitProvider } from "@rainbow-me/rainbowkit";
 
 const config = getDefaultConfig({
   appName: "DM3 delivery service setup helper",
-  projectId: "YOUR_PROJECT_ID",
+  projectId: process.env.REACT_APP_WALLET_CONNECT_PROJECT_ID as string,
   chains: [mainnet, optimism, sepolia],
 });
