+++ src/polyfills.ts
@@ -1,7 +1,4 @@
-//import { Buffer } from 'buffer';
-
 window.global = window.global ?? window;
-//window.Buffer = window.Buffer ?? Buffer;
 window.process = window.process ?? { env: {} }; // Minimal process polyfill
 
 export {};
